half edge data structure and navmesh added
in the navmesh class you will find the fucntion with the name "FindPath" in the NavMesh class where i  implemented the a* algorithm.
when the editor opens look to the console and the function Asearch in the GameSystem class there is an example of home to use the NavMesh class.
issues:
-The GetWalkableNeighbors in the NavMesh class gets only the 3 adjacent triangles by the edges
and i think we need to get neighors by the vertices also.
we need to visualize the nav mesh  and the nodes of the path created by the a* algorithm search to further test the algorithm.